extract data logic from render function
