+++ .gitignore
@@ -9,3 +9,4 @@
 
 # Output of the go coverage tool, specifically when used with LiteIDE
 *.out
+.idea
