+++ webapp/index.html
@@ -64,21 +64,18 @@
       <button data-minutes="1440">24 hours</button>
       <button data-minutes="14400">10 days</button>
     </div>
+    <div>
+      <select id="chart-type">
+        <option value="tau">TauCharts</option>
+        <option value="e">ECharts</option>
+        <option value="any">AnyCharts</option>
+      </select>
+    </div>
     <p id="error"></p>
     <div class="charts">
       <div class="chart" id="top-files"></div>
       <div class="chart" id="top-nodes"></div>
     </div>
-    <div class="charts">
-      <div class="chart" id="tau-top-files"></div>
-      <div class="chart" id="tau-top-nodes"></div>
-    </div>
-    <div class="charts">
-      <div class="chart" id="top-files-e"></div>
-      <div class="chart" id="top-nodes-e"></div>
-    </div>
-    <script src="index.js"></script>
-    <script src="tau.js"></script>
-    <script src="echarts.js"></script>
+    <script type="module" src="index.js"></script>
   </body>
 </html>
